refactor: merge package helper and package generator
